% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Abs_diff_multi}
\alias{Abs_diff_multi}
\title{Plot heatmap of Absolute difference between estimated and real cell type proportions}
\usage{
Abs_diff_multi(prop.real, prop.est, method.name = NULL, title = NULL, ...)
}
\arguments{
\item{prop.real}{a matrix of real cell type proportions}

\item{prop.est}{a matrix or a list of matrices of estimated cell type proportions}

\item{method.name}{vector of the names of estmation methods. Default is NULL and the names will be
generated automatically as 'Est1', 'Est2', ...}

\item{title}{a string of the tile of the plot}
}
\value{
a `ggplot` object with [ggplot2::geom_tile]
}
\description{
Generate a series of heatmap with RMSE and mAD between real and estimated cell type
proportions
}
\seealso{
\code{\link{Prop_comp_multi}}, \code{\link{Eval_multi}},
\code{\link{Scatter_multi}}
}
