% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_notes.R
\name{add_notes}
\alias{add_notes}
\alias{add_notes.m_xtibble}
\title{Add notes to an \code{xtibble} object}
\usage{
add_notes(x, root_note)

\method{add_notes}{m_xtibble}(x, root_note = "C")
}
\arguments{
\item{x}{An \code{\link[=xtibble]{xtibble()}} with \code{\link[=index]{index()}} valid attributes.}

\item{root_note}{A string, the root note (default is "C").}
}
\value{
The \code{\link[=xtibble]{xtibble()}} object with an additional (or updated) column \code{note}.
}
\description{
Add a column with notes to an \code{\link[=xtibble]{xtibble()}} object.
The \code{\link[=xtibble]{xtibble()}} must be ordered as fifths.
A \emph{root note} is required (the note that corresponds to zero in the index).
}
\examples{
start_with_mode(degree = 2) |> add_notes(root_note = "G")
}
\seealso{
Other add_: 
\code{\link{add_roles}()}
}
\concept{add_}
