% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_roles.R
\name{add_roles}
\alias{add_roles}
\alias{add_roles.m_xtibble}
\title{Add roles to an \code{xtibble} object}
\usage{
add_roles(x, ...)

\method{add_roles}{m_xtibble}(x, style = "keep", ...)
}
\arguments{
\item{x}{An \code{\link[=xtibble]{xtibble()}} with \code{\link[=index]{index()}} valid attributes.}

\item{...}{Further arguments passed to or from other methods.}

\item{style}{A string. The style for roles label (when \code{period < 31}):
\itemize{
\item "keep" (default): use 31-TET roles.
\item "collapse_all": all merged roles are renamed.
\item "collapse_simpler": only discarded roles are renamed.
}}
}
\value{
The \code{\link[=xtibble]{xtibble()}} object with an additional (or updated) column \code{role}.
}
\description{
Add a column with roles to an \code{\link[=xtibble]{xtibble()}} object.
Its index must be ordered as fifths.
If the period
}
\examples{
start_with_mode(degree = 7) |> add_roles()
start_with_mode(degree = 7) |> add_roles(style = "collapse_simpler")
}
\seealso{
Other add_: 
\code{\link{add_notes}()}
}
\concept{add_}
