% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Anova_info}
\alias{Anova_info}
\title{ANOVA Test for each gene}
\usage{
Anova_info(eset, non.zero = TRUE, markers = NULL, clusters, samples,
  select.ct = NULL, num.info = 25, ...)
}
\arguments{
\item{eset}{single cell ExpressionSet}

\item{non.zero}{logical, default as TRUE. If TRUE, we only use gene that have non-zero expression}

\item{markers}{vector of characters, default as NULL. If NULL, use all genes in \code{eset}}

\item{clusters}{character, the name of phenoData used as clusters}

\item{samples}{character, the name of phenoData used as samples}

\item{select.ct}{vector of cell types included, default as \code{NULL}. If \code{NULL}, include all cell types in \code{x}}

\item{num.info}{numeric, number of selected gene for each cell type. Default at 25}
}
\value{
a list of
       * F statistics of \code{samples}
       * F statistics of \code{clusters},
       * F statistics of two-way anova
       * selected informative genes: high F statictis for \code{clusters} compare to \code{samples}
}
\description{
This function is to test the two-way significance of cell type specific expression cross samples
}
