% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_12_TET.R
\name{as_12_TET}
\alias{as_12_TET}
\alias{as_12_TET.m_index}
\alias{as_12_TET.m_xtibble}
\title{Set the attributes of an index so that it matches 12-TET}
\usage{
as_12_TET(x, ...)

\method{as_12_TET}{m_index}(x, ...)

\method{as_12_TET}{m_xtibble}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object with modified index.
}
\description{
Set the period to 12 and the unit to 1.
}
\examples{
index(-1:1, period = 31L, unit = 18L) |> as_12_TET()
start_with_some_roles(c("P1", "d5", "A6")) |> as_12_TET()
}
