\name{Batch_adjust}
\alias{Batch_adjust}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Adjusting batch effect.
}
\description{
  Most single cell CRISPR screening data are generated without batch effect. If the data were generated by different batches, this function should be performed to check if there is batch effect and then adjust it.
}
\usage{
Batch_adjust(expression_profile, sample_info_batch, sample_info_gene, plot = TRUE, plot_path = "~/batch_check.pdf")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{expression_profile}{
  A dataframe showing the expression profile after performing cell quality control.
}
  \item{sample_info_batch}{
  A character vector showing the batch of each sample. The vector's column name is the sample name. 
}
  \item{sample_info_gene}{
   A character vector showing the knockout gene of each sample. The vector's column name is the sample name.
}
  \item{plot}{
  FALSE by default. If TURE, plot the graph showing batch effect.
}
  \item{plot_path}{
  The path of the graph you plot. It works only when the parameter "plot" is TRUE.
}
}
\value{
You will get the expression profile after batch effect adjusting.
}
\references{
Manimaran S, Selby HM, Okrah K, Ruberman C, Leek JT, Quackenbush J, Haibe-Kains B, Bravo HC and Johnson WE(2016). “BatchQC: interactive software for evaluating sample and batch effects in genomic data.”_Bioinformatics_. doi: 10.1093/bioinformatics/btw538
Jeffrey T. Leek, W. Evan Johnson, Hilary S. Parker, Elana J. Fertig, Andrew E. Jaffe and John D. Storey (2016). sva: Surrogate Variable Analysis. R package version 3.22.0.
}
\author{
Bin Duan
}
