% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_point.R
\name{break_point}
\alias{break_point}
\alias{break_point.m_index}
\alias{break_point.m_xtibble}
\title{Break the index at a chosen value}
\usage{
break_point(x, ...)

\method{break_point}{m_index}(x, at, ...)

\method{break_point}{m_xtibble}(x, at, ...)
}
\arguments{
\item{x}{An object}

\item{...}{further arguments passed to or from other methods.}

\item{at}{An integer. The breaking point.}
}
\description{
Choose a breaking point for a \emph{circular} index (when the range of values does not exceed the period).
Force numeration to a chosen pattern, by specifying the maximum positive value (next is negative).
}
\examples{
index(0:11, period = 12) |> break_point(at = 5L)
example("xtibble") |> break_point(at = 1L)
}
