% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_chord}
\alias{build_chord}
\title{Build Chord}
\usage{
build_chord(
  root,
  chord = "minor",
  plot = TRUE,
  play = FALSE,
  format_notation = TRUE,
  ...
)
}
\arguments{
\item{root}{String: Root note}

\item{chord}{String: Chord to build. Default = "minor"}

\item{plot}{Logical: If TRUE, plot chord notes using \link{cplot_piano}}

\item{play}{Logical: If TRUE, play chord using \link{play_chord}}

\item{format_notation}{Logical: If TRUE, format notes to include both flats
and sharps to avoid repeating the same letter. e.g. convert c("Gb4", "G4") to
c("F#4", "G4")}

\item{...}{Additional arguments to be passed to \link{play_chord} if
\code{play = TRUE}}
}
\description{
Build Chord
}
\examples{
build_chord("C4", "minor")
build_chord("A4", "sus2", plot = TRUE)
\dontrun{
  build_chord("B4", "sus2", play = TRUE)
}
}
\author{
EDG
}
