% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{CellTotal.df}
\alias{CellTotal.df}
\title{Compare cell type specific total expression (library size) between 2 dataset}
\usage{
CellTotal.df(sce, clusters, samples)
}
\description{
Compare cell type specific total expression (library size) between 2 dataset
}
