% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_roles.R
\name{collapse_roles}
\alias{collapse_roles}
\title{Collapse roles into a lower period}
\usage{
collapse_roles(period = 12, style = c("all", "simpler"))
}
\arguments{
\item{period}{An integer, the new EDO system. 19 or 12.}

\item{style}{A string, either "full" or "simpler"}
}
\value{
An \code{\link[=xtibble]{xtibble()}} object with a new period and new labels for the roles.
}
\description{
Construct new labels for 19 or 12-EDO. Two style are proposed:
\itemize{
\item "all": relabel all merge.
\item "simpler": only relabel the more subtle roles to emphasize the merge,
but keep the merge implicit elsewhere.
}
}
\examples{
collapse_roles(period = 12, style = "all")
collapse_roles(period = 12, style = "simpler")
}
