% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.piano.R
\name{cplot.piano}
\alias{cplot.piano}
\title{Console plot notes}
\usage{
cplot.piano(notes = buildScale("C4", "minor"), blackKey.col = "white")
}
\arguments{
\item{notes}{String, vector: Notes to highlight. Default = \code{buildScale("C4", "minor")}}

\item{blackKey.col}{Color to use for black keys. Default = "white" for use on a dark terminal.
Set to "black" for use on a light terminal.}
}
\description{
Build ASCII plot of a piano with a set of notes highligthed
}
\examples{
cplot.piano(buildScale("B4", "minor"))
}
\author{
Efstathios D. Gennatas
}
