% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot_piano.R
\name{cplot_piano}
\alias{cplot_piano}
\title{Console piano plot}
\usage{
cplot_piano(notes = build_scale("C4", "minor"), blackKey.col = "white")
}
\arguments{
\item{notes}{String, vector: Notes to highlight.
Default = \code{build_scale("C4", "minor")}}

\item{blackKey.col}{Color to use for black keys. Default = "white" for use
on a dark terminal. Set to "black" for use on a light terminal.}
}
\description{
Build an ASCII plot of notes on a piano
}
\examples{
cplot_piano(build_scale("B4", "minor"))
}
\author{
EDG
}
