% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bare.R
\name{create_aeolian_A}
\alias{create_aeolian_A}
\title{Create the aeolian A scale}
\usage{
create_aeolian_A()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} 7x2 with columns \code{idx_r} (\strong{regular half-tones scale}) and \code{note}.
}
\description{
Create the aeolian A scale.
It only builds a regular \code{\link[tibble:tibble]{tibble::tibble()}}.
It purposefully makes no use of anything from this package.
}
\seealso{
Other create_bare: 
\code{\link{create_chromatic}()}
}
\concept{create_bare}
