% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bare.R
\name{create_chromatic}
\alias{create_chromatic}
\title{Create chromatic scales}
\usage{
create_chromatic(flavor = "b")
}
\arguments{
\item{flavor}{A string. Either "b" (default) or "#"/"s". Whether to label the five supplementary notes as sharps or bemols.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} 12 x 2 with columns \code{ìdx_r} (\strong{regular half-tones scale}) and \code{note}.
}
\description{
Create a chromatic scale, either with sharps or bemols.
Because it builds up on \code{\link[=create_aeolian_A]{create_aeolian_A()}}, the chromatic scale will starts on "A".
}
\examples{
create_chromatic(flavor = "#")
}
\seealso{
Other create_bare: 
\code{\link{create_aeolian_A}()}
}
\concept{create_bare}
