% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_notes.R
\name{create_notes}
\alias{create_notes}
\title{Create 31 notes}
\usage{
create_notes()
}
\value{
An \code{\link[=xtibble]{xtibble()}} object with:
\itemize{
\item \code{idx} (period 31, unit 18),
\item \code{note},
\item \code{n_suffix}, number of bemols (negative count) or sharps (positive).
}
}
\description{
Create 31 notes (31-EDO) with "bb", "b", none, "#" and "x" suffix.
"x" is meant as \emph{doubled sharp}.
"Fbb", "Cbb", "Ex", "Bx" are excluded.

A secondary column \code{allowed_root} limit the range of allowed root note from "Gb" to "A#".
This allows to construct -9:9 available notes (19-EDO) for each root note.
}
\examples{
music:::create_notes()
}
