% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Est_sim}
\alias{Est_sim}
\title{Estimate cell type proportions by different methods and evaluate estimations}
\usage{
Est_sim(bulk.eset, sc.eset, p.bulk)
}
\arguments{
\item{p.bulk}{}
}
\value{
a matrix of evaluations
}
\description{
Estimate cell type proportions by different methods and evaluate estimations
}
\details{

}
