% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example}
\alias{example}
\title{Construct tiny examples of objects of class \code{m_index} or \code{m_xtibble}}
\usage{
example(class = NULL, period = 4L, unit = 1L, lock = FALSE)
}
\arguments{
\item{class}{A string. Either "index" or "xtibble".}

\item{period}{An integer. Default is \code{4L}.}

\item{unit}{An integer. Default is \code{1L}.}

\item{lock}{A logical. Default is \code{FALSE}.}
}
\value{
An object of the required class
}
\description{
Construct tiny examples of \code{\link[=index]{index()}} and \code{\link[=xtibble]{xtibble()}} objects.
\itemize{
\item For concise clearer examples in documented function.
\item For the user to get familiar with some fonctions.
}
}
\examples{
example("index")
example("xtibble", period = 7L)
}
