% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_fn.R
\name{formatNotation}
\alias{formatNotation}
\title{Format Notation}
\usage{
formatNotation(notes)
}
\arguments{
\item{notes}{String, vector: Notes to format}
}
\description{
Changes flats to sharps if a note of same letter is already present in the sequence.
e.g. convert the C4 Lydian from:
"C4"  "D4"  "E4"  "Gb4" "G4"  "A4"  "B4"  "C5"
to:
"C4"  "D4"  "E4"  "F#4" "G4"  "A4"  "B4"  "C5"
}
\author{
Efstathios D. Gennatas
}
