% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get internally stored data}
\usage{
get_data(name = NULL)
}
\arguments{
\item{name, }{A string, the name of the table.
If \code{NULL} (default), \code{get_data()} lists available names.}
}
\value{
\itemize{
\item If name is \code{NULL} or invalid, the names of available data.
\item The required data.
}
}
\description{
Proxy function to access preotected data of the package.
}
\examples{
get_data()
get_data(get_data()[1])
}
