% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Get modes of seven or five notes}
\usage{
get_mode(mode_size = 7)
}
\arguments{
\item{mode_size}{An integer. The family of modes. Either 7 or 5.}
}
\value{
An \code{\link[=xtibble]{xtibble()}} object with 3 columns \code{idx}, \code{degree} and \code{name}.
}
\description{
Get all heptatonic or pentatonic modes.
}
\examples{
# All 7-notes modes and initiate index on dorian
get_mode(mode_size = 7)
}
\seealso{
Other mode: 
\code{\link{show_mode}()},
\code{\link{start_with_mode}()}
}
\concept{mode}
