% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_notes.R
\name{get_notes}
\alias{get_notes}
\title{Get notes with zero of the index set to a root note}
\usage{
get_notes(root_note = "C")
}
\arguments{
\item{root_note}{A string, the root note.}
}
\value{
An \code{\link[=xtibble]{xtibble()}} object 31 x 2 with \code{idx} and \code{note}.
}
\description{
Get all 31 notes ordered as fifths, with zero of the index set to a specified root note.
Available root note are any note between "Gb" and "A#".
}
\examples{
get_notes(root_note = "G")
}
