% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roles.R
\name{get_roles}
\alias{get_roles}
\title{Get roles}
\usage{
get_roles()
}
\value{
An \code{\link[=xtibble]{xtibble()}} with 25 roles.
}
\description{
Get roles from internal data and transform them:
\itemize{
\item \code{\link[=index]{index()}} with \code{period = 31}, \code{unit = 18} (fifths);
\item centered on \code{P1}.
}
}
\examples{
get_roles()
}
