% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_upper_tri}
\alias{get_upper_tri}
\title{Get upper triangle matrix}
\usage{
get_upper_tri(cormat)
}
\arguments{
\item{cormat}{square matrix}
}
\value{
upper triangle matrix
}
\description{
Get upper triangle matrix
}
