% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_index.R
\name{index}
\alias{index}
\alias{is_index}
\title{\code{index} vector}
\usage{
index(x = integer(), period = 1L, unit = 1L)

is_index(x)
}
\arguments{
\item{x}{An integer vector
\itemize{
\item For \code{index()}: A integer vector
\item For \code{is_index()}: An object to test.
}}

\item{period}{An integer scalar. The period of the index.}

\item{unit}{An integer scalar. The unit of the index.}
}
\value{
An S3 vector of class \code{m_index}.
}
\description{
Create an integer vector that represents an index with \emph{boosted} features.
}
\details{
\itemize{
\item It is periodic. In the context of the package, it typically has period 31, 19 or 12 (for the 12-TET).
It can also have period 7 or 5 if we want to use these methods for 7-notes or 5-notes sets.
\item It has a transformable \code{unit}. This means the index can represent steps of 1 or any amount of steps.
It can walk by 2 (e.g 0 is 0/12, 1 is 2/12, ...), 3, 4 and so on....
By fifths is the mostly used feature.
}
}
\examples{
index(0:11, period = 12, unit = 7)
}
