% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lock.R
\name{is_locked}
\alias{is_locked}
\title{Test if the \code{m_xtibble} is locked}
\usage{
is_locked(x)
}
\arguments{
\item{x}{An oject of class \code{m_xtibble}.}
}
\value{
A logical
}
\description{
Test if the \code{m_xtibble} is locked
}
