% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Jitter_Est}
\alias{Jitter_Est}
\title{Jitter plot of estimated cell type proportions}
\usage{
Jitter_Est(prop.est, method.name = NULL, title = NULL, ...)
}
\arguments{
\item{prop.est}{a matrix or a list of matrices of estimated cell type proportions}

\item{method.name}{vector of the names of estmation methods. Default is NULL and the names will be
generated automatically as 'Est1', 'Est2', ...}

\item{title}{a string of the tile of the plot}
}
\value{
a 'ggplot' object with [ggplot2::geom_jitter]
}
\description{
Generate jitter plots of estimated cell type proportions cell type by cell type
}
\seealso{
\code{\link{Boxplot_Est}}, \code{\link{Prop_heat_Est}}
}
