\name{MASCOT-package}
\alias{MASCOT-package}
\alias{MASCOT}
\docType{package}
\title{
\packageTitle{MASCOT}
}
\description{
\packageDescription{MASCOT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MASCOT}
\packageIndices{MASCOT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MASCOT}

Maintainer: \packageMaintainer{MASCOT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
