% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{music.iter.ct}
\alias{music.iter.ct}
\title{Scaling bulk data and signature matrix and estimate cell type proportion}
\usage{
music.iter.ct(Y, D, S, Sigma.ct, iter.max = 1000, nu = 1e-04,
  eps = 0.01, centered = FALSE, normalize = FALSE)
}
\arguments{
\item{Y}{vector of bulk tissue expression}

\item{S}{vector of Avg. Library size}

\item{Sigma.ct}{matrix of Subject level variation with cross-cell-type covariance.}

\item{iter.max}{numeric, maximum iteration number. Default is 1000}

\item{nu}{regulation parameter, take care of weight when taking recipical
by default nu = 0.0001.}

\item{eps}{Thredshold of convergence. Default is 0.01.}

\item{centered}{logic, substract avg of Y and D.
Default is FALSE.}

\item{normalize}{logic, divide Y and D by their standard deviation.
Default is FALSE}

\item{X}{matrix, Signature matrix}
}
\value{
a list same as nnls.weight.basic
}
\description{
Scaling bulk data and signature matrix and estimate cell type proportion
}
