% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{music.iter}
\alias{music.iter}
\title{Scaling bulk data and signature matrix and estimate cell type proportion}
\usage{
music.iter(
  Y,
  D,
  S,
  Sigma,
  iter.max = 1000,
  nu = 1e-04,
  eps = 0.01,
  centered = FALSE,
  normalize = FALSE
)
}
\arguments{
\item{Y}{vector of bulk tissue expression}

\item{D}{matrix, Signature matrix}

\item{S}{vector of Avg. Library size}

\item{Sigma}{matrix of subject level variation (gene * cell type)}

\item{iter.max}{numeric, maximum iteration number}

\item{nu}{regulation parameter, when take recipical}

\item{eps}{thredshold of converagence}

\item{centered}{logic, substract avg of Y and D}

\item{normalize}{logic, divide Y and D by their standard deviation}
}
\value{
a list same as nnls.weight.basic
}
\description{
Scaling bulk data and signature matrix and estimate cell type proportion
}
