% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{music_M.theta}
\alias{music_M.theta}
\title{Cross-subject Mean of Relative Abudance}
\usage{
music_M.theta(x, non.zero, markers, clusters, samples, select.ct)
}
\arguments{
\item{x}{ExpressionSet, single cell dataset}

\item{non.zero}{logical, if true, remove all gene with zero expression}

\item{markers}{vector or list of gene names}

\item{clusters}{character, the phenoData used as clusters}

\item{samples}{character,the phenoData used as samples}

\item{select.ct}{vector of cell types included, default as \code{NULL}. If \code{NULL}, include all cell types in \code{x}}
}
\value{
gene by cell type matrix of average relative abundance
}
\description{
This function is for calculating the cross-subject mean of relative abundance for selected cell types.
}
