% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{music_prop.cluster}
\alias{music_prop.cluster}
\title{MuSiC Deconvolution with Clusters}
\usage{
music_prop.cluster(
  bulk.mtx,
  sc.sce,
  group.markers,
  groups,
  clusters,
  samples,
  clusters.type,
  verbose = TRUE,
  iter.max = 1000,
  nu = 1e-04,
  eps = 0.01,
  centered = FALSE,
  normalize = FALSE,
  ...
)
}
\arguments{
\item{bulk.mtx}{Matrix of expression for bulk data}

\item{sc.sce}{SingleCellExperiment for single cell data}

\item{group.markers}{list of gene names. The list include differential expressed genes within groups. 
List name must be the same as `clusters.type`.}

\item{groups}{character, the colData of single cell data used as groups;}

\item{clusters}{character, the colData of single cell dataset used as clusters;}

\item{samples}{character,the colData of single cell dataset used as samples;}

\item{clusters.type}{list of cell types. The list identify groups of similar cell types.}

\item{verbose}{logical, default as TRUE.}

\item{iter.max}{numeric, maximum iteration number}

\item{nu}{regulation parameter, take care of weight when taking reciprocal}

\item{eps}{Threshold of convergence}

\item{centered}{logic, subtract avg of Y and D}

\item{normalize}{logic, divide Y and D by their standard deviation}
}
\value{
matrix of estimated proportions by MuSiC with cluster information.
}
\description{
This function is to calculate the MuSiC deconvolution proportions with clusters
}
\seealso{
\code{\link{music_basis}}; \code{\link{music_prop}}
}
