% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_music_prop.R
\name{music_prop_modified}
\alias{music_prop_modified}
\title{MuSiC Deconvolution}
\usage{
music_prop_modified(bulk.eset, sc.eset = NULL, markers = NULL,
  clusters, samples, select.ct = NULL, ct.cov = FALSE,
  verbose = TRUE, iter.max = 1000, nu = 1e-04, eps = 0.01,
  centered = FALSE, normalize = FALSE, celltype.batching = FALSE,
  celltype.eSet.dir = NULL, ...)
}
\arguments{
\item{bulk.eset}{ExpressionSet for bulk data}

\item{sc.eset}{ExpressionSet for single cell data}

\item{markers}{Vector or list of gene names, default as NULL. If NULL, use
all genes that provided by both bulk and single cell dataset.}

\item{clusters}{Character, the phenoData of single cell dataset used as
clusters;}

\item{samples}{Character,the phenoData of single cell dataset used as
samples;}

\item{select.ct}{Vector of cell types, default as NULL. If NULL, then use all
cell types provided by single cell dataset;}

\item{ct.cov}{Logical. If TRUE, use the covariance across cell types;}

\item{verbose}{Logical, default as TRUE.}

\item{iter.max}{Numeric, maximum iteration number}

\item{nu}{Regulation parameter, take care of weight when taking reciprocal}

\item{eps}{Thredshold of convergence}

\item{centered}{Logic, substract avg of Y and D}

\item{normalize}{Logic, divide Y and D by their standard deviation}

\item{celltype.batching}{Logical. If TRUE, no longer require one
expressionSet for the entire dataset. Instead, dataset can be subsetted by
cell type into a number of expressionSets. I.e. Each expressionSet contains
all samples and all genes commonly expressed (non-zero expression) across
all samples, with only a few cell types from the entire dataset. This works
well for large datasets that have > 50,000 cells/nuclei in total.}

\item{celltype.eSet.dir}{Character. File path to directory containing all
cell-type subsetted expressionSets for an entire dataset. Each
expressionSet should be saved as an .Rds object.}
}
\value{
a list with elements: * Estimates of MuSiC * Estimates of NNLS *
  Weight of MuSiC * r.squared of MuSiC * Variance of MuSiC estimates
}
\description{
This function is to calculate the MuSiC deconvolution proportions
}
\seealso{
\code{\link{music_basis}}
}
