% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{music_S}
\alias{music_S}
\title{Cell type specific library size}
\usage{
music_S(x, non.zero, clusters, samples, select.ct)
}
\arguments{
\item{x}{ExpressionSet, single cell dataset}

\item{non.zero}{logical, if true, remove all gene with zero expression}

\item{clusters}{character, the phenoData used as clusters}

\item{samples}{character,the phenoData used as samples}

\item{select.ct}{vector of cell types included, default as \code{NULL}. If \code{NULL}, include all cell types in \code{x}}
}
\value{
subject by cell type matrix of library
}
\description{
This function is for calculating the cell type specific library size for selected cell types.
}
