% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{music_Sigma.ct}
\alias{music_Sigma.ct}
\title{Cross-subject Corvriance of Relative Abundance}
\usage{
music_Sigma.ct(x, non.zero, markers, clusters, samples, select.ct)
}
\arguments{
\item{x}{ExpressionSet, single cell dataset}

\item{non.zero}{logical, if true, remove all gene with zero expression}

\item{markers}{vector or list of gene names}

\item{clusters}{character, the phenoData used as clusters}

\item{samples}{character,the phenoData used as samples}

\item{select.ct}{vector of cell types included, default as \code{NULL}. If \code{NULL}, include all cell types in \code{x}}
}
\value{
celltype^2 by gene matrix of covariance
}
\description{
This function is for calculating the cross-subject covariance of relative abundance for selected cell types.
}
