% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{music_Theta}
\alias{music_Theta}
\title{Subject and cell type specific relative abundance}
\usage{
music_Theta(x, non.zero = FALSE, clusters, samples, select.ct = NULL)
}
\arguments{
\item{x}{ExpressionSet, single cell dataset}

\item{non.zero}{logical, defualt as F. If true, remove all gene with zero expression}

\item{clusters}{character, the phenoData used as clusters}

\item{samples}{character,the phenoData used as samples}

\item{select.ct}{vector of cell types included, default as \code{NULL}. If \code{NULL}, include all cell types in \code{x}}

\item{markers}{vector or list of gene names}
}
\value{
gene*subject by cell type matrix of relative abundance
}
\description{
This function is for calculating the subject and cell type specific relative abundance for selected cell types.
}
