% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my.log}
\alias{my.log}
\title{log transformation with regulation}
\usage{
my.log(x, nu)
}
\arguments{
\item{x}{numeric}

\item{nu}{regulation}
}
\description{
log transformation with regulation
}
