% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my.rowMeans}
\alias{my.rowMeans}
\title{Calculate Row Means with NA remove}
\usage{
my.rowMeans(x, na.rm = FALSE, dims = 1L)
}
\arguments{
\item{x}{matrix}

\item{na.rm}{default FALSE}
}
\value{
vector of row means
}
\description{
Calculate Row Means with NA remove
}
