% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my.sqrt}
\alias{my.sqrt}
\title{Calculate square root of all values}
\usage{
my.sqrt(x)
}
\arguments{
\item{x}{real number}
}
\value{
sqrt(x)   x >= 0
        -sqrt(-x)  x < 0
}
\description{
Calculate square root of all values
}
