% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note2freq.R
\name{note2freq}
\alias{note2freq}
\title{Convert musical notes to their frequencies}
\usage{
note2freq(note, A4 = 440, default.octave = 4)
}
\arguments{
\item{note}{String: Note(s) to convert to frequencies}

\item{A4}{Float: Frequency for A4. Default = 440}

\item{default.octave}{Integer: If \code{note} is provided without octave number (e.g. "C"), default to this
octave. Default = 4}
}
\description{
Convert musical notes to their frequencies
}
\author{
Efstathios D. Gennatas
}
