% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_music.R
\name{note_distance}
\alias{note_distance}
\title{Note distance in semitones}
\usage{
note_distance(notes)
}
\arguments{
\item{notes}{String, vector: Notes in form \code{c("C4", "Eb4", "Gb4")}}
}
\value{
Vector of length \code{length(notes)} with semitone distances
between successive notes
}
\description{
Calculates note distance in semitones
}
\examples{
note_distance(strings("C4 Eb4 Gb4 Bb4"))
}
\author{
EDG
}
