% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_fn.R
\name{noteDistance}
\alias{noteDistance}
\title{Note distance in semitones}
\usage{
noteDistance(notes)
}
\arguments{
\item{notes}{String, vector: Notes in form \code{c("C4", "Eb4", "Gb4")}}
}
\value{
Vector of length \code{length(notes)} with semitone distances between successive notes
}
\description{
Calculates note distance in semitones
}
\examples{
noteDistance(strings("C4 Eb4 Gb4 Bb4"))
}
\author{
Efstathios D. Gennatas
}
