% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic_slice.R
\name{periodic_slice}
\alias{periodic_slice}
\alias{periodic_slice.m_xtibble}
\title{Slice an \code{xtibble} using index positions}
\usage{
periodic_slice(x, ...)

\method{periodic_slice}{m_xtibble}(x, idx, ...)
}
\arguments{
\item{x}{An \code{\link[=xtibble]{xtibble()}} object.}

\item{...}{Further arguments passed to or from other methods.}

\item{idx}{A vector of integer.}
}
\value{
An \code{\link[=xtibble]{xtibble()}} object.
}
\description{
Slice an \code{\link[=xtibble]{xtibble()}} object, with a list of indices. Importantly:
\itemize{
\item This uses the \code{period} attribute to match indices based on \code{modulo}results.
\item This uses an \emph{inner join}: so this allows pseudo-duplication (for example to select -6:6 of an index with period 12);
It silently ignore missing indices (contrary to a \emph{left join}) so that it allows re-ordering operation by specifying an arbitrary covering range (see example).
}
}
\examples{
tbl <- tibble::tibble(idx = index(0:6, 7L, 1L), name = LETTERS[1:7])
xtibble(tbl, idx_name = "idx") |> periodic_slice(-9:9)

# Re-ordering with a covering range
start_with_mode() |> add_roles() |> periodic_slice(0:11)
}
