% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play_wave}
\alias{play_wave}
\title{Minimal "Polyphonic" Wave Player}
\usage{
play_wave(wave, sample.rate = 44100, plot = FALSE)
}
\arguments{
\item{wave}{Matrix or vector of waveforms. If a matrix, each column should
be a waveform to be played simultaneously}

\item{sample.rate}{Integer: Sample rate. Default = 44100}

\item{plot}{Logical: If TRUE: plot wave using \link{mplot}.}
}
\description{
Play one or more waveforms at the same time using \code{audio::play}
}
\examples{
\dontrun{
play_wave(freq2wave(440))
}
}
\author{
EDG
}
