% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{playWave}
\alias{playWave}
\title{Minimal "Polyphonic" Wave Player}
\usage{
playWave(wave, sample.rate = 44100, plot = FALSE)
}
\arguments{
\item{wave}{Matrix or vector of waveforms. If a matrix, each column should be a waveform to be played
simultaneously}

\item{sample.rate}{Integer: Sample rate. Default = 44100}

\item{plot}{Logical: If TRUE: plot wave using \link{mplot}.}
}
\description{
Play one or more waveforms at the same time using \code{audio::play}
}
\examples{
\dontrun{
playWave(freq2wave(440))
}
}
\author{
Efstathios D. Gennatas
}
