% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCellTotal.two}
\alias{plotCellTotal.two}
\title{Plot the cell type specific library size of 2 single cell datasets}
\usage{
plotCellTotal.two(
  sce1,
  sce2,
  clusters = "cellType",
  samples = "sampleID",
  name1,
  name2
)
}
\arguments{
\item{sce1}{SingleCellExpression of the first single cell dataset}

\item{sce2}{SingleCellExpression of the second single cell dataset}
}
\description{
This function is to plot the cell type specific library size of the common cell types of 2 single cell datasets
}
