% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Prop_comp_multi}
\alias{Prop_comp_multi}
\title{Plot heatmap of real and estimated cell type proportions}
\usage{
Prop_comp_multi(
  prop.real,
  prop.est,
  method.name = NULL,
  title = NULL,
  eval = TRUE,
  ...
)
}
\arguments{
\item{prop.real}{a matrix of real cell type proportions}

\item{prop.est}{a matrix or a list of matrices of estimated cell type proportions}

\item{method.name}{vector of the names of estimation methods. Default is NULL and the names will be
generated automatically as 'Est1', 'Est2', ...}

\item{title}{a string of the title of the plot}

\item{eval}{logical, default as TRUE. If TRUE, pearson correlation evaluation will be printed on the figure.}
}
\value{
a 'ggplot' object with [ggplot2::geom_tile]
}
\description{
Generate heatmaps of real and estimated cell type proportions side by side. Pearson correlation can be calculated and printed.
}
\examples{
# generate random data

}
\seealso{
\code{\link{Abs_diff_multi}}, \code{\link{Eval_multi}},
\code{\link{Scatter_multi}}
}
