% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Prop_heat_Est}
\alias{Prop_heat_Est}
\title{Heatmap of estimated cell type proportions}
\usage{
Prop_heat_Est(prop.est, method.name = NULL, title = NULL, ...)
}
\arguments{
\item{prop.est}{a matrix or a list of matrices of estimated cell type proportions}

\item{method.name}{vector of the names of estmation methods. Default is NULL and the names will be
generated automatically as 'Est1', 'Est2', ...}

\item{title}{a string of the tile of the plot}
}
\value{
a 'ggplot' object with [ggplot2::geom_tile]
}
\description{
Generate heatmap of estimated cell type proportions cell type by cell type
}
\seealso{
\code{\link{Boxplot_Est}}, \code{\link{Jitter_Est}}
}
