% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relative.ab}
\alias{relative.ab}
\title{Calculate relative abundance}
\usage{
relative.ab(X, by.col = TRUE)
}
\arguments{
\item{X}{non-negative matrix for calculate relative abundance}

\item{by.col}{logical, default as TRUE}
}
\description{
Calculate relative abundance
}
