% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Relative_gene_boxplot}
\alias{Relative_gene_boxplot}
\title{Boxplot of relative abundance}
\usage{
Relative_gene_boxplot(
  sc.sce,
  gene.name,
  nu = 10^{
     -10
 },
  marker.id = NULL,
  log.trans = TRUE,
  select.ct = NULL,
  ...
)
}
\arguments{
\item{sc.sce}{ExpressionSet for single cell data}

\item{gene.name}{character, for gene name}

\item{nu}{numeric regulator for log transformation}

\item{marker.id}{numeric indicator of cell type marker genes. Order in the select cell types.
Default is NULL. If NULL, do not print out the cell type.}

\item{log.trans}{logical, default as TRUE. If FALSE, do not take log transformation of relative abundance.}

\item{select.ct}{vector of characters. Default is NULL. If NULL use all cell types provided. Otherwise select
cell types provided.}
}
\value{
a 'ggplot' object with [ggplot2::geom_boxplot]
}
\description{
Generate boxplot of cell type specific relative abundance for each subjects
}
