% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Scatter_multi}
\alias{Scatter_multi}
\title{Plot Scatter plot of real and estimated cell type proportions}
\usage{
Scatter_multi(prop.real, prop.est, method.name = NULL, title = NULL,
  oneline = FALSE, ...)
}
\arguments{
\item{prop.real}{a matrix of real cell type proportions}

\item{prop.est}{a matrix or a list of matrices of estimated cell type proportions}

\item{method.name}{vector of the names of estmation methods. Default is NULL and the names will be
generated automatically as 'Est1', 'Est2', ...}

\item{title}{a string of the tile of the plot}

\item{oneline}{logical, default as FALSE. If TRUE, set \code{nrow = 1} in \code{facet_wrap()}, all plots are in one row.}
}
\value{
a 'ggplot' object with [ggplot2::geom_point]
}
\description{
Generate a series of scatter plot for each cell type between real and estimated cell type
proportions
}
\seealso{
\code{\link{Prop_comp_multi}}, \code{\link{Abs_diff_multi}},
\code{\link{Eval_multi}}
}
