% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_first.R
\name{set_first}
\alias{set_first}
\alias{set_first.m_xtibble}
\title{Set first of the set}
\usage{
set_first(x, rank)

\method{set_first}{m_xtibble}(x, rank = 1L)
}
\arguments{
\item{x}{An \code{\link[=xtibble]{xtibble()}} with \code{\link[=index]{index()}} set to unit 1.}

\item{rank}{An integer. The rank to set first (to set to lowest index value).}
}
\value{
The object with modified index and re-arranged accordingly.
}
\description{
Transform the index so that the \code{i_th} row (when the index is positive) become the first one (with negative lowest value then).
This takes advantage of the periodic feature of the index.
More pragmatically, this computes:
\itemize{
\item any \emph{inversion} of a chord (first inversion with \code{rank = 2} and so on...);
\item any mode of a scale.
}
}
\examples{
c("F", "C", "G") |> start_with_some_notes() |> as_12_TET() |> set_first(rank = 2L)
start_with_mode() |> add_notes() |> set_unit_to_1() |> set_first(rank = 2L)
}
