% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_lock.R
\name{set_lock}
\alias{set_lock}
\title{Lock or unlock an object of class \code{m_xtibble}}
\usage{
set_lock(x, to, silently = FALSE)
}
\arguments{
\item{x}{An object of class \code{m_xtibble}.}

\item{to}{A logical to set the lock to (\code{TRUE} for \emph{locked}).}

\item{silently}{A logical to use the fonction silently in non interactive context.}
}
\value{
Modified \code{x}.
}
\description{
Lock or unlock an object of class \code{m_xtibble}
}
\examples{
example("xtibble") |> set_lock(to = TRUE) |> set_lock(to = FALSE)
}
