% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_period.R
\name{set_period}
\alias{set_period}
\alias{set_period.m_index}
\alias{set_period.m_xtibble}
\title{Set the period to a new value}
\usage{
set_period(x, ...)

\method{set_period}{m_index}(x, to, merge = FALSE, ...)

\method{set_period}{m_xtibble}(x, to, merge = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to or from other methods.}

\item{to}{An integer. The period to set the index to (31, 19 or 12).}

\item{merge}{A logical. Default is FALSE.}
}
\value{
A modified object.
}
\description{
Change the period of an index.
The index must be ordered as fifths (as specified jointly by \code{period} and \code{unit}).

If the new period is lower, there is a possibility of colliding values.
By default they will not be merged and the operation will abort.
In such a case, this usually means the data (notes and/or roles) need some attention.
Merging can be asked but will strip out any other column.

If columns \code{note} and/or \code{role} are present, these richer values will be inherited by the new index.
This is a chosen feature that allows subtle roles definitions even when switching to 12-EDO.
This can be overwritten by calling \code{\link[=add_notes]{add_notes()}} or \code{\link[=add_roles]{add_roles()}} again: both will provide new labels.
}
\examples{
index(0:1, period = 31L, unit = 18L) |> set_period(to = 12L)
start_with_some_roles(c("P1", "d7", "A5")) |> set_period(to = 12L)
start_with_some_roles(c("P1", "d7", "M6")) |> set_period(to = 12, merge = TRUE)
}
