% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_unit.R
\name{set_unit}
\alias{set_unit}
\alias{set_unit.default}
\alias{set_unit.m_index}
\alias{set_unit.m_xtibble}
\title{Set the \code{unit} of an index to a new value}
\usage{
set_unit(x, ...)

\method{set_unit}{default}(x, period, from, to, ...)

\method{set_unit}{m_index}(x, to, ...)

\method{set_unit}{m_xtibble}(x, to, ...)
}
\arguments{
\item{x}{\itemize{
\item A vector of integer.
\item An object.
}}

\item{...}{Further arguments passed to or from other methods.}

\item{period}{An integer, the period of the index.}

\item{from}{An integer, the unit of the current index.}

\item{to}{An integer, the unit to the set the index to.}
}
\value{
\itemize{
\item A vector of integer, the new values of the index.
\item An object, with the new index.
}
}
\description{
Transform the \code{\link[=index]{index()}} to a new unit.
}
