% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_unit_to_1.R
\name{set_unit_to_1}
\alias{set_unit_to_1}
\alias{set_unit_to_1.m_index}
\alias{set_unit_to_1.m_xtibble}
\title{Set unit of an index to one}
\usage{
set_unit_to_1(x)

\method{set_unit_to_1}{m_index}(x)

\method{set_unit_to_1}{m_xtibble}(x)
}
\arguments{
\item{x}{An object.}
}
\value{
The object with a modified index.
}
\description{
Transform the index so that its unit equals one. Values are projected between \code{0} and \code{period-1}.
This wraps \code{\link[=set_unit]{set_unit()}}.
}
\examples{
index(-1:1, period = 12L, unit = 7L) |> set_unit_to_1()
start_with_some_notes(c("C","E","G")) |> set_unit_to_1()
}
