% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_zero.R
\name{set_zero}
\alias{set_zero}
\alias{set_zero.m_index}
\alias{set_zero.m_xtibble}
\title{Set zero on a value of the index itself or from an alternative source}
\usage{
set_zero(x, ...)

\method{set_zero}{m_index}(x, on, ...)

\method{set_zero}{m_xtibble}(x, on, from = NULL, direction = "forward", ...)
}
\arguments{
\item{x}{An object.}

\item{...}{further arguments passed to or from other methods.}

\item{on}{\itemize{
\item An scalar integer if \code{from} is \code{NULL}.
\item A scalar of the same type as \code{from}.
}}

\item{from}{A string, a column name.}

\item{direction, }{A string, either "forward" (default) or "backward".
If \code{from} is supplied, direction used to search \code{on} if it is not unique.}
}
\value{
The modified \code{x} object.
}
\description{
Reset zero of the index on a specified value.

The value \code{on} refers to:
\itemize{
\item the index itself if \code{from} is \code{NULL} (default);
\item a value in the column designated by \code{from} if \code{from} is supplied.
}
}
\examples{
example("index") |> set_zero(on = 1L)
example("xtibble") |> set_zero(on = 2)
example("xtibble") |> set_zero(on = "x2", from = "name")
}
