% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_fn.R
\name{sharps2flats}
\alias{sharps2flats}
\title{Sharps to flats}
\usage{
sharps2flats(notes)
}
\arguments{
\item{notes}{String: Notes to convert to flat notation}
}
\description{
Renames notes to use flats instead of sharps
}
