% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_roles.R
\name{shift_role}
\alias{shift_role}
\alias{augment}
\alias{augment.m_xtibble}
\alias{diminish}
\alias{diminish.m_xtibble}
\title{Augment or diminish roles}
\usage{
shift_role(x, roles, delta)

augment(x, roles)

\method{augment}{m_xtibble}(x, roles)

diminish(x, roles)

\method{diminish}{m_xtibble}(x, roles)
}
\arguments{
\item{x}{An \code{\link[=xtibble]{xtibble()}} object with a column 'role'.}

\item{roles}{A string, scalar or vector. The roles to augment or diminish.}

\item{delta}{An integer. The shift. -7L (diminish) or 7L (augment).}
}
\value{
A new \code{\link[=xtibble]{xtibble()}} with 2 columns:
\itemize{
\item \code{idx} the modified index with attributes set to period 31 and unit 18,
\item \code{role} the modified roles.
}
}
\description{
Augment or diminish roles.
This will drop any other column. Index will be reset with the richest attributes.
}
\examples{
start_with_mode(degree = 6) |> add_roles() |> augment("m3")
start_with_mode(degree = 1) |> add_roles() |> diminish("M7")
}
