% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_with_mode.R
\name{show_mode}
\alias{show_mode}
\title{Show mode}
\usage{
show_mode(mode_size = 7L)
}
\arguments{
\item{mode_size}{An integer, 7 (default) or 5, the required family of modes.}
}
\description{
Show mode
}
\examples{
show_mode(mode_size = 7L)
}
\seealso{
Other mode: 
\code{\link{get_mode}()},
\code{\link{start_with_mode}()}
}
\concept{mode}
