% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_with_mode.R
\name{start_with_mode}
\alias{start_with_mode}
\title{Start with a mode}
\usage{
start_with_mode(mode_size = 7L, degree = 1L)
}
\arguments{
\item{mode_size}{An integer, 7 (default) or 5, the required family of modes.}

\item{degree}{An integer, the degree from the family of modes.}
}
\value{
An \code{\link[=xtibble]{xtibble()}} object with column \code{ìdx} (period = 12; unit = 7), locked.
}
\description{
Start with one of 12 available modes:
\itemize{
\item 7 heptatonic modes;
\item 5 pentatonic modes.
Currently degrees for pentatonic modes are numbered to match heptatonic ones.
The five available degrees are therefore 1, 2, 3, 5 and 6 (and not 1:5).
}
}
\examples{
start_with_mode(mode_size = 7L, degree = 2L)
}
\seealso{
Other mode: 
\code{\link{get_mode}()},
\code{\link{show_mode}()}

Other start_with_: 
\code{\link{start_with_some_notes}()},
\code{\link{start_with_some_roles}()},
\code{\link{start_with_some_values}()}
}
\concept{mode}
\concept{start_with_}
