% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_with_some_notes.R
\name{start_with_some_notes}
\alias{start_with_some_notes}
\title{Start with a set of chosen notes}
\usage{
start_with_some_notes(notes, root_note = "C", idx_name = "idx")
}
\arguments{
\item{notes}{A vector of strings. The set of notes.
Expected format is an uppercase letter between A and G, with the optional suffix "b", "#", "bb" or "x" (e.g. "C", "Eb", "F#"...).}

\item{root_note}{A string. The root note. Not necessary in \code{notes}.}

\item{idx_name}{A string, the column name for the index (default "idx").}
}
\value{
An object \code{m_xtibble} with columns \code{idx} (period 31, unit 18) and \code{notes}.
}
\description{
Create an object of class \code{m_xtibble} from a set of notes.
}
\examples{
start_with_some_notes(c("F","A","Eb"), root_note = "Gb")
}
\seealso{
Other start_with_: 
\code{\link{start_with_mode}()},
\code{\link{start_with_some_roles}()},
\code{\link{start_with_some_values}()}
}
\concept{start_with_}
