% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_with_some_roles.R
\name{start_with_some_roles}
\alias{start_with_some_roles}
\title{Start with a set of chosen roles}
\usage{
start_with_some_roles(roles, idx_name = "idx")
}
\arguments{
\item{roles}{A vector of strings. The set of roles.}

\item{idx_name}{A string, the column name for the index (default "idx").}
}
\value{
An \code{\link[=xtibble]{xtibble()}} with columns \code{idx} (period 31, unit 18) and \code{role}.
}
\description{
Create an \code{\link[=xtibble]{xtibble()}} object from \emph{roles}.
}
\examples{
start_with_some_roles(c("M3", "m7"))
}
\seealso{
Other start_with_: 
\code{\link{start_with_mode}()},
\code{\link{start_with_some_notes}()},
\code{\link{start_with_some_values}()}
}
\concept{start_with_}
