% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_with_some_values.R
\name{start_with_some_values}
\alias{start_with_some_values}
\title{Start with a set of chosen values (of an index)}
\usage{
start_with_some_values(x, period = 31L, unit = 18L, idx_name = "idx")
}
\arguments{
\item{x}{A vector of strings. The set of roles.}

\item{period}{An integer, the period of the \code{\link[=index]{index()}} (default 31L).}

\item{unit}{An integer, the unit of the \code{\link[=index]{index()}} (default 18L).}

\item{idx_name}{A string, the column name for the index (default "idx").}
}
\value{
An \code{\link[=xtibble]{xtibble()}} with one column.
}
\description{
Create an \code{\link[=xtibble]{xtibble()}} object from values of an index.
Higher order helper to construct an \code{\link[=xtibble]{xtibble()}} object and its index with a shorter call.
}
\examples{
# Ionian mode
start_with_some_values((0:6) - 1)
}
\seealso{
Other start_with_: 
\code{\link{start_with_mode}()},
\code{\link{start_with_some_notes}()},
\code{\link{start_with_some_roles}()}
}
\concept{start_with_}
