% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_fn.R
\name{strings}
\alias{strings}
\title{Separate notes into vector of strings}
\usage{
strings(x, sep = " ")
}
\arguments{
\item{x}{String: A single character object which consists of multiple notes separated by \code{sep}
e.g. \code{"C4 Eb4 G4 D5"}}

\item{sep}{String: the character that separates notes in \code{x}. Default = " "}
}
\description{
Convenience function to separate notes into vector of strings
}
\details{
Makes it easy to copy-paste notes into other functions
e.g. \code{playChord(strings("C4 Eb4 G4 D5"))}
}
\examples{
strings("C4 Eb4 Gb4 Bb4")
}
\author{
Efstathios D. Gennatas
}
