% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Twocelltype.Generator}
\alias{Twocelltype.Generator}
\title{Simulate Single cell read counts}
\usage{
Twocelltype.Generator(
  n.sc = 10,
  n.bulk = 100,
  G = 100,
  mu,
  sub.var,
  sort.bulk = FALSE,
  sigma = NULL,
  ...
)
}
\arguments{
\item{n.bulk}{integer, number of bulk subjects.}

\item{G}{integer, number of genes.}

\item{mu}{matrix, 2 by G, the mean expression of genes in each cell types.}

\item{sub.var}{matrix, 2 by G, scale of log-normal distribution for each subjects at genes. Part of them are zeros.}

\item{sort.bulk}{logical, default is FALSE. Sort the cell type proportion.}

\item{N}{integer, number of subjects in total.}
}
\value{
a list with elements
\itemize{
   \item {p.bulk: cell type proportions of generated bulk data;}
   \item {p.sc: cell type proportions of generated single cell data;}
   \item {bulk.mtx: Expression Matrix of bulk data;}
   \item {sc.sce: SingleCellExperiment of single cell data;}
}
}
\description{
Simulate expected library sizes from a log-normal distribution
}
\details{
Generate simulation datasets. n.sc or n.bulk is zero, then we only generate
one dataset and cell type proportions.
}
