% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{w_nnls_S}
\alias{w_nnls_S}
\title{The same as w_nnls_M.theta, except do not use markers and clusters.}
\usage{
w_nnls_S(x, non.zero, clusters, samples, select.ct)
}
\description{
The same as w_nnls_M.theta, except do not use markers and clusters.
}
