% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{w_nnls_Sigma.ct}
\alias{w_nnls_Sigma.ct}
\title{The same as w_nnls_M.theta}
\usage{
w_nnls_Sigma.ct(x, non.zero, markers, clusters, samples, select.ct)
}
\description{
The same as w_nnls_M.theta
}
