% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{w_nnls_Theta}
\alias{w_nnls_Theta}
\title{non.zero defalt as FALSE, other things are the same as w_nnls_M.theta}
\usage{
w_nnls_Theta(x, non.zero = F, clusters, samples, select.ct = NULL)
}
\description{
non.zero defalt as FALSE, other things are the same as w_nnls_M.theta
}
