% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{weight.cal.ct}
\alias{weight.cal.ct}
\title{Calculate weight with cross cell type covariance}
\usage{
weight.cal.ct(Sp, Sigma.ct)
}
\arguments{
\item{Sp}{vector, library size of each cell type}

\item{Sigma.ct}{matrix, cell type^2 by genes. Each columns store the cross cell-type covariance matrix of a gene.}
}
\value{
vector of weight
}
\description{
Calculate weight with cross cell type covariance
}
