% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Weight_cal}
\alias{Weight_cal}
\title{Calculate weight with cross-subject variance for each cell types}
\usage{
Weight_cal(p, M.S, Sigma)
}
\arguments{
\item{p}{vector, cell type proportions}

\item{M.S}{vector of average library size for each cell type}

\item{Sigma}{gene by cell type matrix. Each entry is the cross-subject variance.}
}
\value{
vector of weight
}
\description{
Calculate weight with cross-subject variance for each cell types
}
