% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_xtibble.R
\name{xtibble}
\alias{xtibble}
\alias{is_xtibble}
\title{Create an instance of class \code{m_xtibble}}
\usage{
xtibble(x, idx_name, lock = TRUE)

is_xtibble(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{xtibble()}: A tibble with a \code{\link[=index]{index()}} column.
\item For \code{is_xtibble()}: An object to test.
}}

\item{idx_name}{A string, the name of a column with a \code{\link[=index]{index()}}.}

\item{lock}{A logical, to lock the object and prevent unexpected transformations.}
}
\value{
An object of class \code{m_xtibble}.
}
\description{
This creates a specialized tibble (subclass of a tibble), with \code{\link[=index]{index()}} related features.
}
\examples{
tbl <- tibble::tibble(idx = index(0:2, period = 7, unit = 1), x = LETTERS[1:3])
xtibble(tbl, idx_name = "idx")
}
