\name{Batch_adjust}
\alias{Batch_adjust}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Adjusting batch effect in single cell CRISPR screening dataset.
}
\description{
Most single cell CRISPR screening data are generated without batch effect. If data were generated by different batches, this function can be performed to check if there is batch effect and then adjust it.
}
\usage{
Batch_adjust(expression_profile, batch_information, perturb_information, plot = FALSE, plot_path = "./batch_check.pdf")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{expression_profile}{
  A dataframe showing the expression profile after performing cell quality control,that is, the function of "Cell_qc()" in this package.
}
  \item{batch_information}{
  A character vector showing the batch of each sample. The vector's name is the sample name. 
}
  \item{perturb_information}{
  A character vector showing the perturbation of each sample. The vector's name is the sample name.
}
  \item{plot}{
  FALSE by default. If TRUE, plot the graph showing if there is batch effect.
}
  \item{plot_path}{
  The save path of the graph you plot. It works only when the parameter "plot" is TRUE.
}
}
\value{
A dataframe, the expression profile after batch effect adjusting.
}
\references{
1. Manimaran S, Selby HM, Okrah K, Ruberman C, Leek JT, Quackenbush J, Haibe-Kains B, Bravo HC and Johnson WE(2016). “BatchQC: interactive software for evaluating sample and batch effects in genomic data.”_Bioinformatics_. doi: 10.1093/bioinformatics/btw538
2. Jeffrey T. Leek, W. Evan Johnson, Hilary S. Parker, Elana J. Fertig, Andrew E. Jaffe and John D. Storey (2016). sva: Surrogate Variable Analysis. R package version 3.22.0.
}
\author{
Bin Duan
}
