% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildChord}
\alias{buildChord}
\title{Build Chord}
\usage{
buildChord(root, chord = "minor", play = FALSE, plot = FALSE, ...)
}
\arguments{
\item{root}{String: Root note}

\item{chord}{String: Chord to build. Default = "minor"}

\item{play}{Logical: If TRUE, play chord using \link{playChord}}

\item{plot}{Logical: If TRUE, plot chord notes using \link{cplot.piano}}

\item{...}{Additional arguments to be passed to \link{playChord} if \code{play = TRUE}}
}
\description{
Build chord
}
\details{
For convenience and consistency, all sharps and flats are denoted using flats
}
\examples{
buildChord("C4", "minor")
buildChord("A4", "sus2", plot = TRUE)
\dontrun{
buildChord("B4", "sus2", play = TRUE)
}
}
\author{
Efstathios D. Gennatas
}
