% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{bulk_construct}
\alias{bulk_construct}
\title{Construct artificial bulk tissue expression from single cell data}
\usage{
bulk_construct(eset, clusters, samples, select.ct = NULL)
}
\arguments{
\item{eset}{ExpressionSet, single cell dataset}

\item{clusters}{character, the phenoData used as clusters}

\item{samples}{character, the phenoData used as samples}

\item{select.ct}{vector of cell types included, default as \code{NULL}. If \code{NULL}, include all cell types in \code{x}.
Otherwise, only cells of selected cell types will be used for construction.}
}
\value{
ExpressionSet of constructed bulk tissue, matrix of cell number in each cluster*sample
}
\description{
Artificial bulk tissue expression is generated by adding gene expressions of all cells from
same subjects.
}
