\name{Input_preprocess_10X}
\alias{Input_preprocess_10X}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The preprocess of input data generated with 10X genomics.
}
\description{
For convenience, if the data is generated by 10X genomics, user can use this function to input them. Of course, you can still handle the data to the standard type, then use "Input_preprocess()" function as well.
}
\usage{
Input_preprocess_10X(directory, batch_information = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{directory}{
  This is a directory, including "genes.tsv", "barcodes.tsv", "matrix.mtx", "cbc_gbc_dict.tsv", "cbc_gbc_dict_grna.tsv".
}
  \item{batch_information}{
  NULL by default. A character vector showing the batch of each sample.
}
}
\value{
A list containing "expression profile", "perturb_information" and "batch_information" if it is not NULL.
}
\author{
Bin Duan
}
