\name{Input_preprocess_10X}
\alias{Input_preprocess_10X}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Checking the input data if the data was generated with 10X.
}
\description{
For convenience, if the data is generated by 10X genomics, user can use this function to input them. Of course, you can still handle the data to the standard type, then use "Input_preprocess()" function as well.
}
\usage{
Input_preprocess_10X(directory, sample_info_batch = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{directory}{
  This is a directory, including "genes.tsv", "barcodes.tsv", "matrix.mtx", "cbc_gbc_dict.tsv", "cbc_gbc_dict_grna.tsv".
}
  \item{sample_info_batch}{
  NULL by default. A character vector showing the batch of each sample.
}
}
\value{
A list containing "expression profile", "sample_info_gene", "sample_info_sgRNA" and "sample_info_batch" if it is not NULL.
}
\references{
http://satijalab.org/seurat/
}
\author{
Bin Duan
}
