% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot.R
\name{mplot}
\alias{mplot}
\title{Plot waveform}
\usage{
mplot(
  x,
  type = "l",
  main = NULL,
  legend = TRUE,
  lwd = 1,
  pty = "m",
  bg = "black",
  fg = "gray50",
  col = "cyan",
  col.axis = "gray50",
  col.lab = "gray50",
  col.main = "gray80",
  col.legend = "white",
  tcl = 0.3,
  xaxt = "s",
  yaxt = "s",
  new = FALSE,
  mgp = c(2, 0, 0),
  mar = NULL,
  oma = NULL,
  ...
)
}
\arguments{
\item{x}{Input}

\item{type}{String: "l" for lines, "p" for points. Default = "l"}

\item{main}{String: Plot title}

\item{legend}{Logical: If TRUE, show legends on plot, if \code{x} has column 
names}

\item{lwd}{Float: Line width. Default = 1}

\item{pty}{String: "m" to fill available device space, "s" for square plot. 
Default = "m"}

\item{bg}{Color: background color}

\item{fg}{Color: foreground color}

\item{col}{Color: Point/line color}

\item{col.axis}{Color: Axes' color}

\item{col.lab}{Color: Label color}

\item{col.main}{Color: Title color}

\item{col.legend}{Color: Legend color}

\item{tcl}{The 'tcl' param of par}

\item{xaxt}{The 'xaxt' param of par}

\item{yaxt}{The 'yaxt' param of par}

\item{new}{The 'new' param of par}

\item{mgp}{The 'mgp' param of par}

\item{mar}{Vector, length 4: Margins for \code{par}}

\item{oma}{Vector, length 4: The 'oma' param of par}

\item{...}{Additional parameters to pass to \code{plot}}
}
\description{
Plot waveform
}
\author{
E.D. Gennatas
}
