% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{music.basic}
\alias{music.basic}
\title{Estimate cell type proportion with MuSiC and NNLS}
\usage{
music.basic(Y, X, S, Sigma, iter.max, nu, eps)
}
\arguments{
\item{Y}{vector of bulk tissue expression}

\item{X}{matrix, Signature matrix}

\item{S}{vector of Avg. Library size}

\item{Sigma}{matrix of Subject level variation (gene * cell type)}

\item{iter.max}{numeric, maximum iteration number}

\item{nu}{regulation parameter, take care of weight when taking reciprocal}

\item{eps}{Threshold of convergence}
}
\value{
a list with elements:
\itemize{
  \item {p.nnls: vector of cell type proportions estimated by nnls (add up to 1);}
  \item {q.nnls: vector of original estimation from nnls;}
  \item {fit.nnls: fitted value of Y from nnls;}
  \item {resid.nnls: residual value from nnls;}
  \item {p.weight: vector of cell type proportions estimated by weighted-nnls (add up to 1);}
  \item {q.weight: vector of original estimation from weight-nnls;}
  \item {fit.weight: fitted value of Y from weighted-nnls;}
  \item {resid.weight: residual value from weighted-nnls;}
  \item {weight.gene: weight calculated from weighted-nnls;}
  \item {converge: 'Reach Maxiter', 'Converge at m';}
  \item {rsd: residual value from weighted-nnls transformed by weight;}
  \item {R.squared: R square of weighted-nnls;}
  \item {var.p: variance of weighted-nnls estimated cell type proportions;}
}
}
\description{
Estimate cell type proportion with MuSiC and NNLS
}
