% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{music_basis}
\alias{music_basis}
\title{Prepare Design matrix and Cross-sample Variance for MuSiC Deconvolution}
\usage{
music_basis(x, non.zero = TRUE, markers = NULL, clusters, samples,
  select.ct = NULL, ct.cov = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{ExpressionSet, single cell dataset}

\item{non.zero}{logical, default as TRUE. If true, remove all gene with zero expression.}

\item{markers}{vector or list of gene names. Default as NULL. If NULL, then use all genes provided.}

\item{clusters}{character, the phenoData used as clusters;}

\item{samples}{character,the phenoData used as samples;}

\item{select.ct}{vector of cell types. Default as NULL. If NULL, then use all cell types provided.}

\item{ct.cov}{logical. If TRUE, use the covariance across cell types.}

\item{verbose}{logical, default as TRUE.}
}
\value{
a list of
    * gene by cell type matrix of Design matrix
    * subject by celltype matrix of Library size
    * vector of average library size for each cell type
    * gene by celltype matrix of average relative abudance
    * gene by celltype matrix of cross-sample variation
}
\description{
This function is used for generating cell type specific cross-sample mean and variance for each gene. Cell type specific library size is also calcualted.
}
