% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{music_prop}
\alias{music_prop}
\title{MuSiC Deconvolution}
\usage{
music_prop(bulk.eset, sc.eset, markers = NULL, clusters, samples,
  select.ct = NULL, ct.cov = FALSE, verbose = TRUE, iter.max = 1000,
  nu = 1e-04, eps = 0.01, inter.as.bseq = F, normalize = F, ...)
}
\arguments{
\item{bulk.eset}{ExpressionSet for bulk data}

\item{sc.eset}{ExpressionSet for single cell data}

\item{markers}{vector or list of gene names, default as NULL. If NULL, use all genes that provided by both bulk and single cell dataset.}

\item{clusters}{character, the phenoData of single cell dataset used as clusters;}

\item{samples}{character,the phenoData of single cell dataset used as samples;}

\item{select.ct}{vector of cell types, default as NULL. If NULL, then use all cell types provided by single cell dataset;}

\item{ct.cov}{logical. If TRUE, use the covariance across cell types;}

\item{verbose}{logical, default as TRUE.}

\item{iter.max}{numeric, maximum iteration number}

\item{nu}{regulation parameter, take care of weight when taking recipical}

\item{eps}{Thredshold of convergence}

\item{inter.as.bseq}{logic, substract avg of Y and D as BSEQ-sc did}

\item{normalize}{logic, divide Y and D by their standard deviation}
}
\value{
a list of
        * Estimates of MuSiC
        * Estimates of NNLS
        * Weight of MuSiC
        * r.squared of MuSiC
        * Variance of MuSiC estimates
}
\description{
This function is to calculate the MuSiC deconvolution proportions
}
\seealso{
\code{\link{music_basis}}
}
