% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_fn.R
\name{noteDistance}
\alias{noteDistance}
\title{Note distance in semitones}
\usage{
noteDistance(notes, verbose = TRUE)
}
\arguments{
\item{notes}{String, vector: Notes in form \code{c("C4", "Eb4", "Gb4")}}

\item{verbose}{Logical: If TRUE, print to console}
}
\value{
Vector of length \code{1 - length(notes)} with semitone distances between notes
}
\description{
Calculates note distance in semitones
}
\author{
Efstathios D. Gennatas
}
