% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCellTotal.two}
\alias{plotCellTotal.two}
\title{Plot the cell type specific library size of 2 single cell datasets}
\usage{
plotCellTotal.two(eset1, eset2, cellType, sampleID, name1, name2)
}
\arguments{
\item{eset1}{ExpressionSet of first single cell dataset}

\item{eset2}{ExpressionSet of second single cell dataset}
}
\description{
This function is to plot the cell type specific library size of the common cell types of 2 single cell datasets
}
