% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Prop_convert}
\alias{Prop_convert}
\title{Compare real cell type proportion with estimated cell type proportion; Evaluation}
\usage{
Prop_convert(prop.real, prop.est, method.name = NULL, ...)
}
\arguments{
\item{prop.real, }{real cell type proportions}

\item{prop.est, }{estimated cell type proportion}

\item{method.name}{vector of estimation methods, default as NULL, name will be provided as numbered.}
}
\value{
data.frame of cell type proportion, cell type, subject
}
\description{
Compare real cell type proportion with estimated cell type proportion; Evaluation
}
